Break CollectionLockup out from CollectionListItem
