+++ app/javascript/mastodon/features/collections/components/collection_lockup.module.scss
@@ -1,0 +1,60 @@
+.content {
+  position: relative;
+  flex-grow: 1;
+  display: flex;
+  align-items: center;
+  column-gap: 12px;
+}
+
+.avatarGrid {
+  position: relative;
+  display: grid;
+  grid-template-columns: repeat(2, min-content);
+  gap: 2px;
+
+  &.avatarGridSensitive {
+    .avatar {
+      filter: blur(4px);
+    }
+  }
+}
+
+.avatar {
+  background: var(--color-bg-brand-softest);
+}
+
+.avatarSensitiveBadge {
+  position: absolute;
+  inset: 0;
+  margin: auto;
+  padding: 3px;
+  width: 18px;
+  height: 18px;
+  border-radius: 8px;
+  fill: var(--color-text-primary);
+  background: var(--color-bg-warning-softest);
+}
+
+.link {
+  display: block;
+  font-size: 15px;
+  font-weight: 500;
+  text-decoration: none;
+  color: var(--color-text-primary);
+
+  &:hover {
+    color: var(--color-text-brand);
+  }
+
+  &::after {
+    // Increase clickable area by extending link across parent
+    content: '';
+    position: absolute;
+    inset: 0;
+  }
+}
+
+.info {
+  font-size: 13px;
+  color: var(--color-text-secondary);
+}
